% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_to_table.R
\name{append_to_table}
\alias{append_to_table}
\title{Append sf table to database}
\usage{
append_to_table(table, table_name, CityCode, conn)
}
\arguments{
\item{table}{the sf table to write}

\item{table_name}{the table name as displayed in the DSB}

\item{CityCode}{citycode}

\item{conn}{a connection to the sandbox database}
}
\value{
a message
}
\description{
Append sf table to database
}
