% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_to_table_gsw_pixels.R,
%   R/append_to_table_gws_pixels.R
\name{append_to_table_gsw_pixels}
\alias{append_to_table_gsw_pixels}
\title{Append results of gsw_extract_pixels run on CityCode to the table gsw_pixels in sandbox database
#' @param CityCode citycode}
\usage{
append_to_table_gsw_pixels(CityCode, conn)

append_to_table_gsw_pixels(CityCode, conn)
}
\arguments{
\item{CityCode}{citycode}

\item{conn}{connection}
}
\value{
a message
}
\description{
Append results of gsw_extract_pixels run on CityCode to the table gsw_pixels in sandbox database
#' @param CityCode citycode

Append to glourb table
}
