% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_to_table_gsw_summary.R
\name{append_to_table_gsw_summary}
\alias{append_to_table_gsw_summary}
\title{Append results of gsw_summarise run on CityCode to the table gsw_summary in sandbox database}
\usage{
append_to_table_gsw_summary(CityCode, conn)
}
\arguments{
\item{CityCode}{citycode}

\item{conn}{a connection to the sandbox database}
}
\value{
a message
}
\description{
Append results of gsw_summarise run on CityCode to the table gsw_summary in sandbox database
}
