% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw_extract_pixels_patches.R
\name{extract_values_in_polygon}
\alias{extract_values_in_polygon}
\title{Extract change_raster pixels in polygon and summarise the results}
\usage{
extract_values_in_polygon(polygon, raster_change)
}
\arguments{
\item{polygon}{the sf polygon in which to extract values}

\item{change_raster}{the GSW change intensity raster}
}
\value{
a summary table by type of change ("neg","neutral", or "pos")
}
\description{
Extract change_raster pixels in polygon and summarise the results
}
