% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw_extract_pixels.R
\name{gsw_extract_pixels}
\alias{gsw_extract_pixels}
\title{Extract 1000 pixels of GSW change intensity per zone x reach polygon for a city}
\usage{
gsw_extract_pixels(thisCityCode, dir_images = "data-raw/images")
}
\arguments{
\item{thisCityCode}{citycode}
}
\value{
a table with gsw_pixels results for the city
}
\description{
Extract 1000 pixels of GSW change intensity per zone x reach polygon for a city
}
