% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw_extract_pixels_patches.R
\name{gsw_extract_pixels_patches}
\alias{gsw_extract_pixels_patches}
\title{Extract 1000 pixels of GSW change intensity per zone x reach polygon for a city}
\usage{
gsw_extract_pixels_patches(
  thisCityCode,
  study_areas,
  dir_images = "data-raw/images"
)
}
\arguments{
\item{thisCityCode}{citycode}

\item{study_areas}{the sf polygons describing reach and zone per city}

\item{dir_images}{the directory containing GSW images}
}
\value{
a table with gsw_pixels results for the city
}
\description{
Extract 1000 pixels of GSW change intensity per zone x reach polygon for a city
}
