% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw_summarise.R
\name{gsw_summarise}
\alias{gsw_summarise}
\title{Summarise results from the 1000 pixels randomly extracted for each zone x reach of a particular city}
\usage{
gsw_summarise(thisCityCode, average_entropy, conn)
}
\arguments{
\item{thisCityCode}{citycode}

\item{average_entropy}{the average entropy (above which entropy is considered high i.e. change is considered diffuse).}
}
\value{
a message
}
\description{
Summarise results from the 1000 pixels randomly extracted for each zone x reach of a particular city
}
