% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw_extract_pixels_patches.R
\name{gsw_summarise_patches}
\alias{gsw_summarise_patches}
\title{get gsw_patches_summary results for the city}
\usage{
gsw_summarise_patches(thisCityCode, conn)
}
\arguments{
\item{thisCityCode}{citycode}

\item{conn}{connection}
}
\value{
a table with gsw_patches_summary results for the city
}
\description{
get gsw_patches_summary results for the city
}
