% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw_summary.R
\name{gsw_summary}
\alias{gsw_summary}
\title{Extract 1000 pixels of GSW change intensity per zone x reach polygon for a city}
\usage{
gsw_summary(city, conn)
}
\arguments{
\item{city}{city name}

\item{conn}{connection to sandbox database}
}
\value{
a table with gsw_summary results for the city
}
\description{
Extract 1000 pixels of GSW change intensity per zone x reach polygon for a city
}
