% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw_summary.R
\name{gsw_summary_plot}
\alias{gsw_summary_plot}
\title{Plot gsw_summary}
\usage{
gsw_summary_plot(gws_summary_result)
}
\arguments{
\item{gsw_summary_result}{result of gsw_summary(city)}
}
\value{
a table with gsw_pixels results for the city
}
\description{
Plot gsw_summary
}
